Updated machine gen code for long and long long ints

+++ machine_gen_code/lookupBestHandPrimes.h
@@ -4,9 +4,9 @@
 
 #include <vector>
 
-std::vector<int> lookupBestHandPrimes(int AP1, int AP2, int AP3, int AP4, int AP5, int AP6, int AP7,\
-                                      int FP1, int FP2, int FP3, int FP4, int FP5, int FP6, int FP7,\
-                                      int SP1, int SP2, int SP3, int SP4, int SP5, int SP6, int SP7);
+std::vector<long long int> lookupBestHandPrimes(int AP1, int AP2, int AP3, int AP4, int AP5, int AP6, int AP7,\
+                                           int FP1, int FP2, int FP3, int FP4, int FP5, int FP6, int FP7,\
+                                           int SP1, int SP2, int SP3, int SP4, int SP5, int SP6, int SP7);
 
 #endif
